### 2025-08-18 #257 @amyheather
lintquarto: Package for running linters, static type checkers and code analysis tools on python code in quarto (.qmd) files.
Submitting Author: (@amyheather)
Package Name: **lintquarto**
One-Line Description of Package: **Package for running linters, static type checkers and code analysis tools on python code in quarto (.qmd) files.**
Repository Link:  **https://github.com/lintquarto/lintquarto**
Version submitted:  **v0.9.0**
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

**This package allows you to run linters, static type checkers, and code analysis tools on Python code embedded in Quarto (.qmd) files. Normally, these tools can't analyse code in Quarto files as they're meant for standard Python scripts, while Quarto files mix code with text and other formats. This package solves that by extracting the Python code from Quarto files with appropriate preservation of line numbers in the code so that the analysis tools can understand and evaluate it correctly.**

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

**While data validation and testing typically focuses on verifying the accuracy and reliability of data, code validation through linting (as well as checks using static type checkers and code analysis tools) is an important upstream step in scientific workflows. Ensuring code standardisation and quality helps improve clarity and maintainability, and may prevent the introduction of errors or inconsistencies into data processing pipelines. By enabling linting of python code within Quarto Markdown files, this package helps maintain high standards for both code and the data it produces or manipulates, supporting reproducible and reliable scientific research.**

**The target audience is scientists, researchers, and analysts who use Quarto Markdown files to document and share computational analyses. It supports scientific workflows by ensuring that embedded Python code is robust, maintainable, and free of common errors.**

**While tools like [nbqa](https://github.com/nbQA-dev/nbQA) allow running linters on Jupyter notebooks, there are currently no Python packages that support linters, static type checkers and code analysis tools to run on python code within .qmd files. This package fills that gap by specifically supporting Quarto Markdown.**

N/A
