+++ .github/workflows/lint.yaml
@@ -18,7 +18,7 @@ jobs:
         with:
           python-version: '3.13'
           cache: 'pip'
-      - run: pip install -r requirements-test.txt
+      - run: pip install -r requirements-dev.txt
 
       - name: Lint package
         run: |
