### 2025-08-19 cea64da
docs(readme): add some more metrics badges (conda download, github stars + forks, github commits + release)
