### 2025-08-19 3598dd9
docs(readme): remove commit activity badge
