### 2025-08-20 09e6024
fix(converter): don't end in_chunk_options until first code line (inc! comments). Add test case to check this. (#59)
