### 2025-08-20 d1df072
fix(process_qmd): actually fix so it is just replacing the base qmd filename with the base py filename (#57)
