+++ src/lintquarto/__main__.py
@@ -6,6 +6,7 @@ Acknowledgements
 Some parts of this module were generated by and adapted from Perplexity.
 """
 
+import os
 import subprocess
 import sys
 from pathlib import Path
@@ -85,14 +86,18 @@ def process_qmd(
         check=False
     )
 
+    # Get the base filename from the full file paths
+    qmd_filename = str(os.path.basename(qmd_path))
+    py_filename = str(os.path.basename(py_file.name))
+
     # Replace all references to the .py file with the .qmd file
-    result.stdout = result.stdout.replace(str(py_file.name), str(qmd_file))
+    result.stdout = result.stdout.replace(py_filename, qmd_filename)
     print(result.stdout, end="")
 
     # If there is an error - which will include some linter outputs that get
     # classed as errors - then also replace `.py` and then print
     if result.stderr:
-        result.stderr = result.stderr.replace(str(py_file.name), str(qmd_file))
+        result.stderr = result.stderr.replace(py_filename, qmd_filename)
         print(result.stderr, file=sys.stderr)
 
     # Remove temporary .py file unless keep_temp_files is set
