### 2025-08-20 a223260
refactor(converter): reduce complexity and repetition in process_qmd
