### 2025-08-20 lintquarto/lintquarto#61 @amyheather
`pyOpenSci` review
Opened issue: https://github.com/pyOpenSci/software-submission/issues/257
