### 2025-08-21 #257 @eliotwrobson
@amyheather thanks for your submission! This seems like a very useful tool. We're going through some policy discussions on the linked issue (https://github.com/pyOpenSci/software-peer-review/issues/331). Once those are resolved, I can do the final EiC checks. I'll see if I can get to some basic steps of this before then, so you don't have to wait as long before getting actionable feedback 👍🏽
