### 2025-08-21 #257 @amyheather
Thank you @eliotwrobson! The use of LLMs is indeed a hot topic at the moment. I hope the group is able to come to a conclusion that still allows my package to be reviewed.

It was interesting to read through the comments so far on https://github.com/pyOpenSci/software-peer-review/issues/331. My take on it is that they are another tool in the "coding toolkit", alongside things like stack overflow, package documentation, and other resources. I've started using them here and there in an iterative way (e.g. help troubleshoot/identify issues, refine/clarify wording in docstrings, refactoring suggestions/better ways of structuring than I might have initially planned). For me, it's not about having an entire block of code written for me, but rather for targeted bits of support.

For that reason, it would be difficult - particularly in retrospect - to point to specific places in the code and say, "this line was generated / adapted from code by an LLM" - in the same way that it would be difficult to identify which were shaped and guided by looking at other sources like package documentation and so on - as it is all iteratively developed with help of looking at lots of resources. Hence, the more "blanket" statement in the README of just acknowledging them generally as having been used during the coding process.

Look forward to hearing where the decision ends - and regardless of the outcome, wanted to thank you all! The resources on the pyOpenSci website are great, and really helped me to build and improve the package.
