### 2025-08-21 4f3f79a
chore(acknowledgements): redo main module docstring + remove individual module messages to just have one overarching statement in README, for consistency / to cover all / reduce repetition
