+++ src/lintquarto/args.py
@@ -1,10 +1,4 @@
-"""
-CustomArgumentParser.
-
-Acknowledgements
-----------------
-Some parts of this module were generated by and adapted from Perplexity.
-"""
+"""CustomArgumentParser."""
 
 import argparse
 import sys
