+++ docs/pages/helpers.py
@@ -1,10 +1,4 @@
-"""
-Helper functions for the tools pages in the user guide.
-
-Acknowledgements
-----------------
-Some parts of this module were generated by and adapted from Perplexity.
-"""
+"""Helper functions for the tools pages in the user guide."""
 
 from typing import Optional
 
