+++ src/lintquarto/converter.py
@@ -1,10 +1,4 @@
-"""
-Convert .qmd file to python file.
-
-Acknowledgements
-----------------
-Some parts of this module were generated by and adapted from Perplexity.
-"""
+"""Convert .qmd file to python file."""
 
 import re
 import warnings
