### 2025-09-01 73aa568
docs(readme): based on pyopensci discussions, adding some more specific details on how AI was used to the README
