### 2025-09-24 7426cb7
feat(include): add handling for quarto include syntax (#65)
