### 2025-09-24 718d2b1
test(include): add test for quarto include syntax fix (#65)
