### 2025-09-24 2d75eca
feat(#<<): add handling for #<< (#63)
