### 2025-09-24 1bfed61
test(#<<): add test for new #<< handling (#63)
