### 2025-09-24 00fe1ac
fix(lint_package): attempt to fix non-exit issue (#67) - adding back in lint issue as test case!
