### 2025-09-24 d16be8c
fix(lint_docs): same fix for non-exit issue (#67) + fixing flake8 warning again
