+++ lint_docs.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+errors=0
+
 # ----------------------------------------------------------------------------
 # Run lintquarto on .qmd files in docs/
 # ----------------------------------------------------------------------------
@@ -12,6 +14,7 @@ LINTERS="ruff flake8 pylint radon-cc vulture pydoclint mypy"
 EXCLUDE="docs/pages/api docs/pages/tools/examples"
 
 lintquarto -l $LINTERS -p docs --exclude $EXCLUDE
+(( errors += $? ))
 
 # ----------------------------------------------------------------------------
 # Run linters on .py files in docs/
@@ -26,22 +29,34 @@ PYFILES=$(find docs -type d -name ".*" -prune -false -o -type f -name "*.py" -pr
 
 echo "Running ruff check..."
 ruff check $PYFILES
+(( errors += $? ))
 
 # Ignore type hint-related warnings (so just arise with pydoclint)
 echo "Running flake8..."
 flake8 $PYFILES --ignore DOC
+(( errors += $? ))
 
 echo "Running pylint..."
 pylint $PYFILES
+(( errors += $? ))
 
 echo "Running radon cc..."
 radon cc $PYFILES
+(( errors += $? ))
 
 echo "Running vulture..."
 vulture $PYFILES vulture/whitelist.py
+(( errors += $? ))
 
 echo "Running pydoclint..."
 pydoclint $PYFILES
+(( errors += $? ))
 
 echo "Running mypy..."
 mypy $PYFILES
+(( errors += $? ))
+
+if [ "$errors" -ne 0 ]; then
+    echo "One or more linting commands failed."
+    exit 1
+fi
