### 2025-09-25 f057ae7
fix(process_line): correct detection of python code chunk start
Now allows space between ``` and {python}, and allows in-line chunk options e.g. {python, echo=FALSE} (#64), and doesn't false-positive include invalid ```{python}```
