### 2025-09-25 0ba5666
test(process_line): add tests for new corrections (#64), and correct test (needing \n after {python})
