### 2025-09-25 c608b94
chore(release): update for 0.6.0 (init, citation.cff, changelog)
