+++ src/lintquarto/__init__.py
@@ -1,3 +1,3 @@
 """lintquarto"""
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
