+++ CITATION.cff
@@ -25,5 +25,5 @@ keywords:
   - static-type-checking
   - quartodoc
 license: MIT
-version: 0.5.0
-date-released: '2025-08-18'
+version: 0.6.0
+date-released: '2025-09-25'
