### 2025-09-25 2b1f3b7
docs(contributing): correct/clarify conda update instructions
