### 2025-09-25 0a9de1c
docs(cli): making cli interface usage/options a little clearer
