### 2025-09-26 01ce1ff
ci(precommit): add back in linting pre-commit
