### 2025-09-29 350479f
docs(contributing): removed emojis
