+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project are documented.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Dates formatted as YYYY-MM-DD as per [ISO standard](https://www.iso.org/iso-8601-date-and-time-format.html).
 
+## Unreleased
+
+TBC
+
 ## v0.6.0 - 2025-09-25
 
 ### Added
