### 2025-10-02 e88fe6f
ci(branch): add extra step to check for exact match
