+++ .github/workflows/branch-enforcement.yml
@@ -1,4 +1,4 @@
-# Source: https://github.com/marketplace/actions/branch-enforcement
+# Adapted from: https://github.com/marketplace/actions/branch-enforcement
 
 name: Branch Enforcement
 
@@ -20,18 +20,12 @@ jobs:
       BASE_REF: ${{ github.base_ref }}
 
     steps:
-      # Policy:
-      #   - Any branch can be merged into dev (e.g. feature branches)
-      #   - dev can be merged into main
-      - name: Enforce Merge Policy
-        id: enforce-policy
-        uses: ncalteen/branch-enforcement@v2
-        with:
-          policy: |
-            *:dev
-            dev:main
-          head_ref: ${{ env.HEAD_REF }}
-          base_ref: ${{ env.BASE_REF }}
+
+      - name: Confirm exact dev->main only
+        if: env.BASE_REF == 'main' && env.HEAD_REF != 'dev'
+        run: |
+          echo "Only 'dev' branch can be merged into 'main'"
+          exit 1
 
       # This step should not run if the branch policy is not met.
       - name: Policy Passed
