### 2025-10-03 9f24af4
docs(exclude): correct exclude command (spaces not commas)
