+++ docs/pages/get_started.qmd
@@ -89,7 +89,7 @@ Lint all files in current directory (using `ruff`):
 * Excluding a specific file `analysis/test.qmd`.
 
 ```{.bash}
-lintquarto -l ruff -p . -e examples,ignore
+lintquarto -l ruff -p . -e examples ignore
 ```
 
 ```{.bash}
