### 2025-10-03 b987453
ci(precommit): edit pre-commit + bash so only run on staged changes
