### 2025-11-04 998b46a
ci(precommit): temporarily stash unstaged changes when run
