+++ .pre-commit-config.yaml
@@ -7,9 +7,11 @@ repos:
         language: system
         pass_filenames: true
         files: \.py$
+        stages: [commit]
       - id: lint-docs
         name: Lint Documentation
         entry: ./lint_docs.sh
         language: system
         pass_filenames: true
         files: ^docs/.*\.(py|qmd)$
+        stages: [commit]
