### 2025-11-04 401e029
ci(precommit): remove stages: [commit] (does anyway, did nothing adding lines)
