### 2025-11-06 4ddcc4b
test(runtime_checkable): add test detecting E302 error for function with protocol (#70)
