### 2025-11-06 876ce0a
chore(runtime_checkable): rename protocol -> decorator
