### 2025-12-04 #257 @eliotwrobson
@amyheather sorry for the wait here! We've had some things moving around behind the scenes, but I think we're _almost_ done with an initial iteration of the policy https://github.com/pyOpenSci/pyopensci.github.io/pull/734 (part of the reason for the slowdown was the length of the comments here).

I will double check but I think the policy is solidified enough that we can go ahead with the review. Thank you again for bearing with us through everything and hopefully we can get things moving soon.
