### 2025-12-05 ec15c6d
rm(branch-enforcement): blocking coverage badge... also decided no longer want this (#82)
