+++ .github/workflows/branch-enforcement.yml
@@ -1,33 +1,0 @@
-# Adapted from: https://github.com/marketplace/actions/branch-enforcement
-
-name: Branch Enforcement
-
-on:
-  pull_request:
-
-jobs:
-  check-branch:
-    name: Check Branches
-    runs-on: ubuntu-latest
-
-    permissions:
-      statuses: write
-
-    env:
-      # The branch being merged.
-      HEAD_REF: ${{ github.head_ref }}
-      # The target branch.
-      BASE_REF: ${{ github.base_ref }}
-
-    steps:
-
-      - name: Confirm exact dev->main only
-        if: env.BASE_REF == 'main' && env.HEAD_REF != 'dev'
-        run: |
-          echo "Only 'dev' branch can be merged into 'main'"
-          exit 1
-
-      # This step should not run if the branch policy is not met.
-      - name: Policy Passed
-        id: passed
-        run: echo "This PR passes the merge policy!"
