### 2025-12-05 fd4a80f
chore(0.7.0): update changelog, citation, readme and __init__ for release
