### 2026-01-08 08ebdb5
docs(get_started): correct styling of conda install command
