### 2026-01-20 7420149
feat/ci(joss): add skeleton JOSS paper + paper build action
