+++ paper.md
@@ -1,0 +1,48 @@
+---
+title: 'lintquarto: a package for running linters, static type checkers and code analysis tools on python code in quarto .qmd files'
+tags:
+  - Python
+  - Quarto
+  - lint
+authors:
+  - name: Amy Heather
+    orcid: 0000-0002-6596-3479
+    equal-contrib: true
+    affiliation: "1"
+affiliations:
+ - name: University of Exeter, England
+   index: 1
+date: 20 January 2026
+bibliography: paper.bib
+---
+
+# Summary
+<!--A description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.-->
+
+TODO
+
+# Statement of need
+<!-- A section that clearly illustrates the research purpose of the software and places it in the context of related work. This should clearly state what problems the software is designed to solve, who the target audience is, and its relation to other work.-->
+
+TODO
+
+# State of the field
+<!--A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear “build vs. contribute” justification explaining your unique scholarly contribution and why existing alternatives are insufficient.-->
+
+TODO
+
+# Software design
+<!--An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application. This should demonstrate meaningful design thinking beyond a superficial code structure description.-->
+
+TODO
+
+# Research impact statement
+<!--Evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals). The evidence should be compelling and specific, not aspirational.-->
+
+TODO
+
+# AI usage disclosure
+
+TODO
+
+# References
