### 2026-01-20 4e697e2
chore(joss): remove paper.bib from paper.md
