### 2026-01-20 0a52b12
ci(draft-pdf): edit so only runs on changes to paper.md
