### 2026-01-20 5b35268
docs(readme): add mention of PyOpenSci, JOSS paper + JOSS GitHub action
