### 2026-01-20 09cdf64
docs(joss): first draft JOSS
