+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'lintquarto: a package for running linters, static type checkers and code analysis tools on python code in quarto .qmd files'
+title: 'lintquarto: a package for running linters, static type checkers and code analysis tools on python code in quarto markdown files'
 tags:
   - Python
   - Quarto
@@ -15,33 +15,89 @@ affiliations:
 date: 20 January 2026
 ---
 
+<!--
+Example paper: https://joss.readthedocs.io/en/latest/example_paper.html
+
+(Optional) create a metadata file describing your software and include it in your repository. We provide a script that automates the generation of this metadata: https://gist.github.com/arfon/478b2ed49e11f984d6fb
+
+https://joss.readthedocs.io/en/latest/paper.html
+ The paper should be between 750-1750 words. 
+-->
+
 # Summary
 <!--A description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.-->
 
-TODO
+`lintquarto` enables existing Python linters, static type checkers, and code analysis tools to run on Python code in Quarto Markdown files. Quarto has become widely adopted in research for creating websites, papers, reports, and documentation that integrate code, narrative, and outputs However, standard Python code quality tools cannot natively run on Quarto's `.qmd` files. `lintquarto` bridges this gap, allowing researchers to maintain Python code quality standards in their Quarto documents.
 
 # Statement of need
 <!-- A section that clearly illustrates the research purpose of the software and places it in the context of related work. This should clearly state what problems the software is designed to solve, who the target audience is, and its relation to other work.-->
 
-TODO
+Researchers can embed executable Python code in Quarto documents using two formats: Jupyter notebooks (`.ipynb`) or Quarto Markdown (`.qmd`). `.qmd` is the standard Quarto format. It integrates better with version control because code and text are stored as plain text, and outputs are generated separately during rendering, not stored in the source file. In contrast, `.ipynb` stores outputs and metadata in JSON, making diffs harder to read.
+
+Python researchers using Quarto Markdown face a critical gap: code quality tools cannot natively run on `.qmd` files. Code quality tools include linters (which identify style violations and potential bugs), static type checkers (which verify type annotations), and code analysis tools (which detect dead code or measure complexity). These tools are essential for maintaining high code standards and ensuring research code is understandable, maintainable, and reliable.
 
 # State of the field
 <!--A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear “build vs. contribute” justification explaining your unique scholarly contribution and why existing alternatives are insufficient.-->
 
-TODO
+For Jupyter notebooks, `nbqa` is the stablished solution, wrapping Python linters and formatters so they can run on `.ipynb` files (CITE). However, researchers preferring Quarto Markdown have no equivalent Python tool.
+
+For R, well-established solutions exist: `lintr` supports code quality checking across Quarto Markdown, R Markdown and plain R files (CITE).
+
+`lintquarto` closes this gap by enabling established Python code quality tools to run on Quarto Markdown files.
 
 # Software design
 <!--An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application. This should demonstrate meaningful design thinking beyond a superficial code structure description.-->
 
-TODO
+From the command line, users specify which code quality tools to run, which paths to include, and which to exclude. Supported tools at the time of publication are detailed in Table []{label="supported_tools"}. `lintquarto` identifies all relevant `.qmd` files and processes these in turn for each code quality tool.
+
[…]
