### 2026-01-20 63f1bf8
docs(joss): add citations
