+++ paper.bib
@@ -1,0 +1,156 @@
+
+@misc{python_code_quality_authority_flake8_2025,
+	title = {flake8},
+	url = {https://github.com/PyCQA/flake8},
+	abstract = {flake8 is a python tool that glues together pycodestyle, pyflakes, mccabe, and third-party plugins to check the style and quality of some python code.},
+	urldate = {2026-01-20},
+	author = {{Python Code Quality Authority}},
+	year = {2025},
+	note = {original-date: 2014-09-13T17:06:24Z},
+	keywords = {complexity-analysis, flake8, linter, linter-flake8, pep8, python, python3, static-analysis, static-code-analysis, style-guide, styleguide, stylelint},
+}
+
+@misc{python_code_quality_authority_pycodestyle_2025,
+	title = {pycodestyle},
+	url = {https://github.com/PyCQA/pycodestyle},
+	abstract = {Simple Python style checker in one Python file},
+	urldate = {2026-01-20},
+	author = {{Python Code Quality Authority}},
+	year = {2025},
+	note = {original-date: 2009-10-02T00:43:37Z},
+	keywords = {flake8-plugin, linter-flake8, linter-plugin, pep8, python, style-guide, styleguide},
+}
+
+@misc{pydoclint_contributors_pydoclint_2025,
+	title = {pydoclint},
+	copyright = {MIT},
+	url = {https://github.com/jsh9/pydoclint},
+	abstract = {A very fast Python docstring linter},
+	urldate = {2026-01-20},
+	author = {{pydoclint contributors}},
+	year = {2025},
+	note = {original-date: 2023-05-12T01:39:32Z},
+	keywords = {docstring, docstring-checker, documantation, documentation-tool, flake8, flake8-extension, flake8-extensions, flake8-plugin, flake8-plugins, linter, python-documentation},
+}
+
+@misc{python_code_quality_authority_pyflakes_2025,
+	title = {pyflakes},
+	copyright = {MIT},
+	url = {https://github.com/PyCQA/pyflakes},
+	abstract = {A simple program which checks Python source files for errors},
+	urldate = {2026-01-20},
+	author = {{Python Code Quality Authority}},
+	year = {2025},
+	note = {original-date: 2014-04-07T06:19:46Z},
+	keywords = {hacktoberfest, linter, python},
+}
+
+@misc{pylint_contributors_pylint_2026,
+	title = {Pylint},
+	copyright = {GPL-2.0},
+	url = {https://github.com/pylint-dev/pylint},
+	abstract = {It's not just a linter that annoys you!},
+	urldate = {2026-01-20},
+	author = {{Pylint contributors}},
+	month = jan,
+	year = {2026},
+	note = {original-date: 2015-12-09T05:40:02Z},
+}
+
+@misc{astral_ruff_2026,
+	title = {Ruff},
+	copyright = {MIT},
+	url = {https://github.com/astral-sh/ruff},
+	abstract = {An extremely fast Python linter and code formatter, written in Rust.},
+	urldate = {2026-01-20},
+	author = {{Astral}},
+	year = {2026},
+	note = {original-date: 2022-08-09T17:17:44Z},
+	keywords = {linter, pep8, python, python3, ruff, rust, rustpython, static-analysis, static-code-analysis, style-guide, styleguide},
+}
+
+@misc{seipp_jendrikseippvulture_2026,
+	title = {jendrikseipp/vulture},
+	copyright = {MIT},
+	url = {https://github.com/jendrikseipp/vulture},
+	abstract = {Find dead Python code},
+	urldate = {2026-01-20},
+	author = {Seipp, Jendrik},
+	month = jan,
+	year = {2026},
[…]
