### 2026-01-20 3e1796f
docs(joss): revise/rewrite
