+++ paper.md
@@ -16,84 +16,58 @@ date: 20 January 2026
 bibliography: paper.bib
 ---
 
-<!--
-Example paper: https://joss.readthedocs.io/en/latest/example_paper.html
-
-(Optional) create a metadata file describing your software and include it in your repository. We provide a script that automates the generation of this metadata: https://gist.github.com/arfon/478b2ed49e11f984d6fb
-
-https://joss.readthedocs.io/en/latest/paper.html
- The paper should be between 750-1750 words. 
--->
-
 # Summary
-<!--A description of the high-level functionality and purpose of the software for a diverse, non-specialist audience.-->
 
-`lintquarto` enables existing Python linters, static type checkers, and code analysis tools to run on Python code in Quarto Markdown files. Quarto [@allaire_quarto_2022] has become widely adopted in research for creating websites, papers, reports, and documentation that integrate code, narrative, and outputs However, standard Python code quality tools cannot natively run on Quarto's `.qmd` files. `lintquarto` bridges this gap, allowing researchers to maintain Python code quality standards in their Quarto documents.
+`lintquarto` enables researchers to apply Python linters, static type checkers, and code analysis tools to Python code embedded within Quarto Markdown (`.qmd`) files. Quarto [@allaire_quarto_2022] has become widely adopted in research for creating websites, papers, reports, and documentation that integrate code, narrative, and outputs However, standard Python code quality tools cannot natively run on Quarto's `.qmd` format. `lintquarto` bridges this gap, enabling researchers to maintain Python code quality standards in their Quarto documents.
 
 # Statement of need
-<!-- A section that clearly illustrates the research purpose of the software and places it in the context of related work. This should clearly state what problems the software is designed to solve, who the target audience is, and its relation to other work.-->
 
 Researchers can embed executable Python code in Quarto documents using two formats: Jupyter notebooks (`.ipynb`) or Quarto Markdown (`.qmd`). `.qmd` is the standard Quarto format. It integrates better with version control because code and text are stored as plain text, and outputs are generated separately during rendering, not stored in the source file. In contrast, `.ipynb` stores outputs and metadata in JSON, making diffs harder to read.
 
 Python researchers using Quarto Markdown face a critical gap: code quality tools cannot natively run on `.qmd` files. Code quality tools include linters (which identify style violations and potential bugs), static type checkers (which verify type annotations), and code analysis tools (which detect dead code or measure complexity). These tools are essential for maintaining high code standards and ensuring research code is understandable, maintainable, and reliable.
 
 # State of the field
-<!--A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear “build vs. contribute” justification explaining your unique scholarly contribution and why existing alternatives are insufficient.-->
 
-For Jupyter notebooks, `nbqa` is the stablished solution, wrapping Python linters and formatters so they can run on `.ipynb` files (CITE). However, researchers preferring Quarto Markdown have no equivalent Python tool.
+For Jupyter notebooks, `nbqa` is the etablished solution, wrapping Python linters and formatters so they can run on `.ipynb` files [@nbqa_contributors_nbqa_2025].
 
-For R, well-established solutions exist: `lintr` supports code quality checking across Quarto Markdown, R Markdown and plain R files (CITE).
+For R, the `lintr` packages supports code quality checking across Quarto Markdown, R Markdown and plain R files [@hester_static_2025].
 
-`lintquarto` closes this gap by enabling established Python code quality tools to run on Quarto Markdown files.
[…]
