+++ paper.bib
@@ -154,3 +154,32 @@
 	doi = {10.5281/zenodo.5960048},
 	file = {Snapshot:/home/amy/Zotero/storage/CJWIZKUS/5960048.html:text/html},
 }
+
+@misc{nbqa_contributors_nbqa_2025,
+	title = {{nbQA}},
+	copyright = {MIT},
+	url = {https://github.com/nbQA-dev/nbQA},
+	abstract = {Run ruff, isort, pyupgrade, mypy, pylint, flake8, and more on Jupyter Notebooks},
+	urldate = {2026-01-20},
+	publisher = {nbQA-dev},
+	author = {{nbQA contributors}},
+	year = {2025},
+	note = {original-date: 2020-07-11T15:30:54Z},
+	keywords = {black, codequality, doctest, flake8, isort, jupyter-notebook, lint, mypy, pre-commit, pre-commit-hook, pylint, python, pyupgrade, ruff, yapf},
+}
+
+@article{hester_static_2025,
+	title = {Static {Code} {Analysis} for {R}},
+	volume = {10},
+	copyright = {http://creativecommons.org/licenses/by/4.0/},
+	issn = {2475-9066},
+	url = {https://joss.theoj.org/papers/10.21105/joss.07240},
+	doi = {10.21105/joss.07240},
+	number = {108},
+	urldate = {2026-01-20},
+	journal = {Journal of Open Source Software},
+	author = {Hester, Jim and Angly, Florent and Chirico, Michael and Hyde, Russ and Kun, Ren and Patil, Indrajeet and Rosenstock, Alexander},
+	month = apr,
+	year = {2025},
+	pages = {7240},
+}
