### 2026-01-20 9ff6a05
refactor(joss): move paper to inst/, add codemeta.json, update readme contributing and action
