### 2026-01-20 #257 @amyheather
**Edit:** Updated to now please automatically submit to the Journal of Open Source Software following PyOpenSci review. Meets all four JOSS checks.
