### 2026-01-22 e1252e6
docs(joss): captialise Quarto/Python, say "Quarto Markdown" instead of qmd, make code quality tool definitions a list so easier to read, remove repetitive sentence at end of state of field
