+++ inst/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'lintquarto: a package for running linters, static type checkers and code analysis tools on python code in quarto markdown files'
+title: 'lintquarto: a package for running linters, static type checkers and code analysis tools on Python code in Quarto Markdown files'
 tags:
   - Python
   - Quarto
@@ -18,13 +18,19 @@ bibliography: paper.bib
 
 # Summary
 
-`lintquarto` enables researchers to apply Python linters, static type checkers, and code analysis tools to Python code embedded within Quarto Markdown (`.qmd`) files. Quarto [@allaire_quarto_2022] has become widely adopted in research for creating websites, papers, reports, and documentation that integrate code, narrative, and outputs However, standard Python code quality tools cannot natively run on Quarto's `.qmd` format. `lintquarto` bridges this gap, enabling researchers to maintain Python code quality standards in their Quarto documents.
+`lintquarto` enables researchers to apply Python linters, static type checkers, and code analysis tools to Python code embedded within Quarto Markdown (`.qmd`) files. Quarto [@allaire_quarto_2022] has become widely adopted in research for creating websites, papers, reports, and documentation that integrate code, narrative, and outputs However, standard Python code quality tools cannot natively run on Quarto Markdown files. `lintquarto` bridges this gap, enabling researchers to maintain Python code quality standards in their Quarto documents.
 
 # Statement of need
 
-Researchers can embed executable Python code in Quarto documents using two formats: Jupyter notebooks (`.ipynb`) or Quarto Markdown (`.qmd`). `.qmd` is the standard Quarto format. It integrates better with version control because code and text are stored as plain text, and outputs are generated separately during rendering, not stored in the source file. In contrast, `.ipynb` stores outputs and metadata in JSON, making diffs harder to read.
+Researchers can embed executable Python code in Quarto documents using two formats: Jupyter notebooks (`.ipynb`) or Quarto Markdown. Quarto Markdown is the standard Quarto format. It integrates better with version control because code and text are stored as plain text, and outputs are generated separately during rendering, not stored in the source file. In contrast, `.ipynb` stores outputs and metadata in JSON, making diffs harder to read.
 
-Python researchers using Quarto Markdown face a critical gap: code quality tools cannot natively run on `.qmd` files. Code quality tools include linters (which identify style violations and potential bugs), static type checkers (which verify type annotations), and code analysis tools (which detect dead code or measure complexity). These tools are essential for maintaining high code standards and ensuring research code is understandable, maintainable, and reliable.
+Python researchers using Quarto Markdown face a critical gap: code quality tools cannot natively run on Quarto Markdown files. Code quality tools include:
+
+* Linters - identify style violations and potential bug.
+* Static type checkers - verify type annotations.
+* Code analysis tools - detect dead code or measure complexity.
+
+These tools are essential for maintaining high code standards and ensuring research code is understandable, maintainable, and reliable.
 
 # State of the field
 
@@ -32,17 +38,18 @@ For Jupyter notebooks, `nbqa` is the etablished solution, wrapping Python linter
 
 For R, the `lintr` packages supports code quality checking across Quarto Markdown, R Markdown and plain R files [@hester_static_2025].
 
-However, researchers preferring Quarto Markdown have no equivalent Python tool. `lintquarto` closes this gap by enabling established Python code quality tools to run on Quarto Markdown files.
+However, researchers preferring Quarto Markdown have no equivalent Python tool.
 
 # Software design
 
 The command-line interface for `lintquarto` accepts three primary arguments: the code quality tools to run, the paths to exclude, and paths to exclude.
 
[…]
