### 2026-01-22 0cf6d29
docs(joss): correct vulture bibtex
