+++ inst/paper.bib
@@ -69,17 +69,16 @@
 	keywords = {linter, pep8, python, python3, static-analysis, static-code-analysis, style-guide, styleguide, ruff, rust, rustpython},
 }
 
-@misc{seipp_jendrikseippvulture_2026,
-	title = {jendrikseipp/vulture},
+@misc{vulture_contributors_vulture_2026,
+	title = {vulture},
 	copyright = {MIT},
 	url = {https://github.com/jendrikseipp/vulture},
 	abstract = {Find dead Python code},
 	urldate = {2026-01-20},
-	author = {Seipp, Jendrik},
-	month = jan,
+	author = {{vulture contributors}},
 	year = {2026},
 	note = {original-date: 2017-03-06T08:00:18Z},
-	keywords = {python, dead-code-removal},
+	keywords = {dead-code-removal, python},
 }
 
 @misc{python_mypy_2026,
