### 2026-01-22 fdbe3ab
refactor(vulture): move from linter to code analysis
