+++ docs/_quarto.yml
@@ -48,7 +48,6 @@ website:
             - pages/tools/pyflakes.qmd
             - pages/tools/pylint.qmd
             - pages/tools/ruff.qmd
-            - pages/tools/vulture.qmd
         - section: "**Static type checkers**"
           contents:
             - pages/tools/mypy.qmd
@@ -57,6 +56,7 @@ website:
             - pages/tools/pytype.qmd
         - section: "**Code analysis tools**"
           contents:
+            - pages/tools/vulture.qmd
             - pages/tools/radon.qmd
 
 metadata-files:
