### 2026-01-22 6ff507e
docs(joss): correct vulture citation
