### 2026-02-04 cea0af2
feat(eval): add methods in QmdToPyConverter that detect eval setting from YAML front matter and from cell chunk options (#88)
