### 2026-02-04 6af0295
test(eval): add tests for the new eval methods (#88)
