### 2026-02-04 da6bbd3
refactor(eval): save YAML front matter default as attribute (#88)
