### 2026-02-04 b4d9a12
refactor/feat(eval): add logic to skip chunks when eval false (#88)
