### 2026-02-05 3227bec
test(eval): add tests for eval control of chunks kept/removed (#88)
