+++ tests/test_converter.py
@@ -55,14 +55,19 @@ def test_non_python_chunk_is_commented(linter):
 # 2. Conversion of active python chunks
 # =============================================================================
 
-def remove_noqa(lines):
+def remove_noqa(lines: list[str]) -> list[str]:
     """
     Helper to remove # noqa comments from expected output
 
     Parameters
     ----------
-    lines : list of str
+    lines : list[str]
         Lines of text (expected output)
+
+    Returns
+    -------
+    list[str]
+        Lines with any trailing '  noqa: ...' comments removed.
     """
     return [
         line.split("  # noqa")[0] if "  # noqa" in line
@@ -573,5 +578,252 @@ def test_parse_yaml_front_matter_invalid_yaml():
 def test_parse_chunk_eval_option(line, expected):
     """Unit: parse_chunk_eval_option parses boolean eval values."""
     converter = QmdToPyConverter(linter="flake8")
-    result = converter.parse_chunk_eval_option(line)
-    assert result is expected
+    converter.parse_chunk_eval_option(line)
+    assert converter.current_chunk_eval is expected
+
+
+# =============================================================================
+# 8. Integration: eval controls which chunks are kept
+# =============================================================================
+
+def _convert(lines):
+    conv = QmdToPyConverter(linter="flake8")
+    return conv.convert(lines)
+
+
+def test_no_yaml_all_chunks_linted_by_default():
+    """Integration: no YAML → all python chunks linted (kept)."""
+    qmd = [
+        "Some text\n",
+        "```{python}\n",
+        "x = 1\n",
+        "```\n",
+        "\n",
+        "```{python}\n",
+        "y = 2\n",
+        "```\n",
+    ]
+    py = _convert(qmd)
+    assert any("x = 1" in line for line in py)
+    assert any("y = 2" in line for line in py)
+    assert len(py) == len(qmd)
+
+
+def test_yaml_eval_false_all_chunks_skipped_by_default():
+    """Integration: YAML eval: false → all chunks skipped unless overridden."""
+    qmd = [
+        "---\n",
+        "execute:\n",
+        "  eval: false\n",
+        "---\n",
+        "\n",
+        "```{python}\n",
+        "x = 1\n",
+        "```\n",
+    ]
+    py = _convert(qmd)
+    assert "x = 1" not in py
+    # still preserving line count
+    assert len(py) == len(qmd)
+
+
+def test_chunk_eval_true_overrides_yaml_false():
+    """Integration: #| eval: true keeps chunk even if YAML eval: false."""
+    qmd = [
+        "---\n",
+        "execute:\n",
+        "  eval: false\n",
+        "---\n",
+        "\n",
+        "```{python}\n",
+        "#| eval: true\n",
+        "x = 1\n",
+        "```\n",
+    ]
+    py = _convert(qmd)
+    assert any("x = 1" in line for line in py)
+
+
+def test_chunk_eval_false_overrides_yaml_true():
+    """Integration: #| eval: false skips chunk even if YAML eval: true."""
+    qmd = [
+        "---\n",
+        "execute:\n",
+        "  eval: true\n",
+        "---\n",
+        "\n",
+        "```{python}\n",
+        "#| eval: false\n",
+        "x = 1\n",
+        "```\n",
+    ]
+    py = _convert(qmd)
+    assert "x = 1" not in py
+
+
+def test_chunk_eval_false_without_yaml():
+    """Integration: #| eval: false skips chunk when no YAML present."""
+    qmd = [
+        "```{python}\n",
+        "#| eval: false\n",
+        "x = 1\n",
[…]
