### 2026-02-05 e449242
fix(eval): correct converter logic for eval (#88)
