### 2026-02-05 a89aa97
chore(dependencies): add yaml to pyproject and types-pyyaml to contributing.md
