### 2026-02-10 4c03507
build(env): switch from virtualenv/conda to uv
