### 2026-02-10 5356f67
build(env): upgrade python + uv run instructions
