### 2026-02-10 f285aeb
fix(uv): fix griffe versions that work
