### 2026-02-11 07ecfe1
docs(contributing): update package build instructions to uv
