### 2026-02-11 4acf25d
chore(v0.8.0): update __init__, README, CITATION.cff and CHANGELOG.md
