+++ src/lintquarto/__init__.py
@@ -1,3 +1,3 @@
 """lintquarto"""
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
