+++ CITATION.cff
@@ -25,5 +25,5 @@ keywords:
   - static-type-checking
   - quartodoc
 license: MIT
-version: 0.7.0
-date-released: '2025-12-05'
+version: 0.8.0
+date-released: '2026-02-11'
