### 2026-02-11 3a13ef2
fix(ci): forgot to update with uv 😂 attempted switch from requirements to uv in docs.yaml
