### 2026-02-11 e95099b
fix(ci): attempt to fix docs.yaml - didn't use uv for render
