+++ .github/workflows/docs.yaml
@@ -6,11 +6,17 @@ on:
     branches: main
   workflow_dispatch:
 
+env:
+  UV_SYSTEM_PYTHON: "1"
+
 jobs:
   build-deploy:
     runs-on: ubuntu-latest
     permissions:
       contents: write
+    env:
+      UV_SYSTEM_PYTHON: "1"
+
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
@@ -23,7 +29,7 @@ jobs:
         with:
           python-version-file: ".python-version"
 
-      - name: Install uv (with cache)
+      - name: Install uv
         uses: astral-sh/setup-uv@v7
         with:
           enable-cache: true
