### 2026-02-11 aa77c7e
fix(ci): attempt to fix docs - didn't use uv for render
