### 2026-02-11 a29522e
fix(ci): attempt to fix docs - didn't use uv for render
