+++ .github/workflows/docs.yaml
@@ -36,16 +36,13 @@ jobs:
         working-directory: docs
         run: uv run quartodoc build
 
-      - name: Discover uv venv Python
-        id: venv
-        run: |
-          echo "PY=$(uv run which python)" >> "$GITHUB_OUTPUT"
+      - name: Quarto render
+        working-directory: docs
+        run: uv run quarto render
 
-      - name: Render and Publish
-        uses: quarto-dev/quarto-actions/publish@v2
-        with:
-          target: gh-pages
-          path: docs
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          QUARTO_PYTHON: ${{ steps.venv.outputs.PY }}
+      - name: Quarto publish to gh-pages
+        working-directory: docs
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          uv run quarto publish gh-pages --no-prompt --no-browser
