### 2026-02-11 b69ceee
fix(ci): fix lint to use uv
