+++ .github/workflows/lint.yaml
@@ -13,17 +13,23 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
       
-      - name: Install python and dependencies
+      - name: Set up Python from .python-version
         uses: actions/setup-python@v4
         with:
-          python-version: '3.13'
-          cache: 'pip'
-      - run: pip install -r requirements-dev.txt
+          python-version-file: ".python-version"
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v7
+        with:
+          enable-cache: true
+
+      - name: Sync project from uv.lock
+        run: uv sync --locked --all-extras --dev
 
       - name: Lint package
         run: |
-            bash lint_package.sh
+            uv run bash lint_package.sh
       
       - name: Lint documentation
         run: |
-            bash lint_docs.sh
+            uv run bash lint_docs.sh
