### 2026-02-11 1f54391
feat(ci): run on pull request for docs, lint and tests (avoiding issue I had just now, of merging to main without realising they would fail)
