+++ .github/workflows/tests.yaml
@@ -4,6 +4,8 @@ run-name: Run tests
 on:
   push:
     branches: [main]
+  pull_request:
+    branches: main
   workflow_dispatch:
 
 jobs:
