+++ .github/workflows/lint.yaml
@@ -4,6 +4,8 @@ run-name: Lint package and documentation
 on:
   push:
     branches: main
+  pull_request:
+    branches: main
   workflow_dispatch:
 
 jobs:
