### 2026-02-11 dbccd36
chore(v0.8.0): update changelog
