### 2026-02-11 c7f6bc3
fix(ci): only upload/commit coverage badge for push (not pull)
