### 2026-02-11 63b810e
feat(ci): add new action to build package, and publish (if release) (updated CONTRIBUTING + CHANGELOG accordingly)
