### 2026-02-24 #257 @kysolvik
Hi @amyheather 👋

My name is Kylen, I'm going to be jumping in to serve as the editor for `lintquarto`'s review. Excited to get started!

Everything looks ready to go thanks to you, @eliotwrobson, and @ab93. Next step is finding reviewers. I'll let you know as soon as we've assigned them. In the meantime, let me know if you have any questions!
