### 2026-03-03 #257 @kysolvik
## Editor comments

Hi @KandiSanjana and @isabelizimm! Thanks so much for volunteering to review `lintquarto`!

Please get in touch with any questions or concerns! Your review is due: March 31st, 2026
---

Reviewers: @KandiSanjana, @isabelizimm
Due date: March 31st, 2026
