### 2026-04-01 #257 @isabelizimm
Hi there! This is a fantastic package. Everything is well organized, and you clearly have a strong attention to detail. This is a package that is hard to write examples for, so the fact you added download buttons + file examples + `lintquarto` examples does not go unnoticed!

Full disclosure, I work at Posit, the company that created Quarto. I don't work on the Quarto team, but have been an occasional contributor as well as HUGE Quarto fan for many years. This package fixes real pain points that I have felt as a user. Thank you so much for your great work! 👾

- [X] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [X] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [X] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

**NOTE**: Some of these notes are non-blocking to the review and are more of a "hey, I'm someone who really loves your package and want to give you use feedback about this" type of comment. I added 🌱 to denote these types of comments. I want to be very clear they are **NON-BLOCKING this `pyOpenSci` review and should be treated as optional.** If you're interested, I'm happy to open up issues on the `lintquarto` repo!

Usage:

- 🌱  [Quarto does have a tree-sitter](https://github.com/quarto-dev/q2/tree/main/crates/tree-sitter-qmd/tree-sitter-markdown) to find Python chunks to avoid using regex. It'd be simpler and more robust, especially for nonstandard Python chunks. Happy to take this one to the pyOpenSci Slack or email or elsewhere and lend a hand for this too!

- Quarto has valid, non-executable Python code noted by `python` code chunks rather than `{python}`. It's worth noting somewhere that only executable code will be validated.
- 🌱 Continuing from above, it could be nice to lint non-executable cells, perhaps in a `--strict` mode or something
- It is a little confusing that the docs say ["Ruff runs as a code formatter, ..." ](https://lintquarto.github.io/lintquarto/pages/tools/ruff.html), but `lintquarto` actually does _not_ run the formatter at all. I think it's worth specifically calling out that `ruff` will ONLY be used as a linter.
- Missing installation instructions for development version of the package, eg, `pip install -e '[all]'`
- `CONTRIBUTING.md` some links not hyperlinked when rendered.
- [References in `paper.md` are blank](https://github.com/lintquarto/lintquarto/blob/main/inst/paper.md), but it looks like some papers were referenced? I am not a JOSS expert, so not super sure on this.

Maintenance and misc:

- The [`lint_package.sh` file](https://github.com/lintquarto/lintquarto/blob/main/lint_package.sh) that lints in CI (and recommended for use in the CONTRIBUTING guide) has potentially redundant checks. You could probably just use ruff rather than ruff+flake8+pylint+radon+vulture. This file does not seem to be used for tests, so I don't think you'd need all these tools?
- Same comment as above, but for the `lint_docs.sh` file
- The `requirements-test.txt` file looks as though it could be refactored as a dependency group in the `pyproject.toml` file, if desired
- For documenting related projects, you might look at [panache](https://github.com/jolars/panache). I think a big differentiator here is that `lintquarto` is a pure Python implementation
