### 2026-05-06 #257 @amyheather
@isabelizimm I have implemented your suggestions. Thank you again for the time you spent on this - I really appreciate it! The only one I haven't done is `tree‑sitter-markdown` - after trying it out, I think I'd like to keep lintquarto's current regex-based chunk detection for now, since it already handles the Quarto Python chunks I need and avoids another dependency. I appreciate the potential benefits around more structured parsing and robustness to edge cases, but for now I'd prefer to keep the implementation simpler and lighter, if that sounds alright? I'm still open to revisiting tree‑sitter - and please do say if you feel that it would be worth pursuing! It is something new to me, so appreciate I may not be considering all angles here! ☺️ Trying to find the right balance between robustness, and complexity and dependencies.

Also @kysolvik - please can I ask the relevant next steps for the review process? Thank you.
